Renamed to match NPM package.
