+++ @google-cloud/index.d.ts
