+++ @google-cloud/storage/index.d.ts
